Refactor: extract pdfsandwich OCR to helper runPdfsandwich
